Check if the domain is in our Azure DNS

+++ move-domains-to-afd.sh
@@ -13,8 +13,8 @@ AFD_ROUTINGRULES="$AFD-routingrule httptohttps"
 KV="MY-KV"
 KV_ID=$(az keyvault list --resource-group $RG  | jq -r '[.[].id]|join("")')
 OLD_FRONTENDS=$(az network front-door frontend-endpoint list --resource-group $RG --front-door-name $AFD | jq -r '[.[].name]|join(" ")' )
-OUTPUT="json" # Change to "none" to get less output
-
+DNS_ZONES=$(az network dns zone list --resource-group $DNS_RG --query '[].name' | jq -r '.|join(" ")')
+OUTPUT="none"
 
 echo -e "\nUPDATING AZURE DNS"
 snames=$(az keyvault certificate list --vault-name $KV | jq -r '[.[].name]|join(" ")')
@@ -22,25 +22,28 @@ for SECRET_NAME in $snames; do
 	CUSTOMDOMAINS=$(az keyvault certificate show --vault-name $KV --name $SECRET_NAME | jq -r '.. | objects | select(.subjectAlternativeNames).subjectAlternativeNames.dnsNames |join(" ")')
 
 	for DOMAIN in  $CUSTOMDOMAINS; do
-
 		echo -e "\nUpdating $DOMAIN"
-		ZONE=$(echo $DOMAIN | rev | cut -d. -f1-2 | rev) # NOTE, I assume domain.tld, if you have domain.co.tld you have to change this.
-		if [[ $ZONE != $DOMAIN ]]; then
-			HOST=$(echo $DOMAIN | cut -d. -f1)
-			echo -e "\tHOST: finding record type for $HOST in zone $ZONE"
-			record_type=$(az network dns record-set list --resource-group $DNS_RG --zone-name $ZONE --query "[?name=='$HOST'].type" | jq -r '.|join("")|split("/")[-1]' )
-			echo -e "\tRecord type: $record_type"
-			if [[ "A" == $record_type ]]; then
-				echo -e "\tDELETE A-record"
-				az network dns record-set a delete --resource-group $DNS_RG --zone-name $ZONE --name $HOST --yes --output $OUTPUT
+		ZONE=$(echo $DOMAIN | rev | cut -d. -f1-2 | rev)
+
+		# If the domain in the certificate is in our Azure DNS, modify it.
+		if [[ $DNS_ZONES =~ (^|[[:space:]])$ZONE($|[[:space:]]) ]]; then
+			if [[ $ZONE != $DOMAIN ]]; then
+				HOST=$(echo $DOMAIN | cut -d. -f1)
+				echo -e "\tHOST: finding record type for $HOST in zone $ZONE"
+				record_type=$(az network dns record-set list --resource-group $DNS_RG --zone-name $ZONE --query "[?name=='$HOST'].type" | jq -r '.|join("")|split("/")[-1]' )
+				echo -e "\tRecord type: $record_type"
+				if [[ "A" == $record_type ]]; then
+					echo -e "\tDELETE A-record"
+					az network dns record-set a delete --resource-group $DNS_RG --zone-name $ZONE --name $HOST --yes --output $OUTPUT
+				fi
+				echo -e "\tCreate CNAME $HOST -> $AFD_HOST"
+				az network dns record-set cname set-record --resource-group $DNS_RG --zone-name $ZONE --record-set-name $HOST --cname $AFD_HOST --output $OUTPUT
+			else
+				echo -e "\tpoint @ to $AFD_HOST"
+				az network dns record-set a update --resource-group $DNS_RG --zone-name $ZONE --name "@"  --target-resource $AFD_ID --output $OUTPUT
+				echo -e "\tAdd CNAME afdverify -> afdverify.$AFD_HOST"
+				az network dns record-set cname set-record --resource-group $DNS_RG --zone-name $ZONE --record-set-name "afdverify" --cname "afdverify.${AFD_HOST}" --output $OUTPUT
 			fi
-			echo -e "\tCreate CNAME $HOST -> $AFD_HOST"
-			az network dns record-set cname set-record --resource-group $DNS_RG --zone-name $ZONE --record-set-name $HOST --cname $AFD_HOST --output $OUTPUT
-		else
-			echo -e "\tpoint @ to $AFD_HOST"
-			az network dns record-set a update --resource-group $DNS_RG --zone-name $ZONE --name "@"  --target-resource $AFD_ID --output $OUTPUT
-			echo -e "\tAdd CNAME afdverify -> afdverify.$AFD_HOST"
-			az network dns record-set cname set-record --resource-group $DNS_RG --zone-name $ZONE --record-set-name "afdverify" --cname "afdverify.${AFD_HOST}" --output $OUTPUT
 		fi
 	done
 done
